move the search code to app.handler
