+++ auth/auth.handler.js
@@ -46,25 +46,7 @@ const login = async (req, res) => {
         res.status(500).json({success: "false", error:"Error logging in"})
     }
 }
-const search = async (req, res) => {
-    const {username, email} = req.body
-    try{
-        const [user] = await find(username, email)
-        return res.status(200).json({
-            success: "true",
-            message: "Search success",
-            data: {
-                id: user.id,
-                username: user.username,
-                email: user.email,
-                created_at:user.created_at
-            }
-        })
-    }catch(error){
-        console.error(error)
-        res.status(500).json({success: "false", error:"Error searching user"})
-    }
-}
+
 module.exports={
-    register, login, search
+    register, login
 }
