+++ test.rest
@@ -20,14 +20,13 @@ POST http://localhost:4000/search HTTP/1.1
 Content-Type: application/json
 
 {
-    "username":"msdn",
-    "password":"jirolupat",
-    "email":"nidu672@gmail.com"
+    "username":"ach",
+    "email":"ach@gmail.com"
 }
 ###
 GET http://localhost:4000/users HTTP/1.1
 ###
-POST http://localhost:4000/users/gUhjgOy7Tn HTTP/1.1
+PUT http://localhost:4000/users/gUhjgOy7Tn HTTP/1.1
 Content-Type: application/json
 
 {
